+++ app/lizard-nxt.js
@@ -78,3 +78,11 @@ angular.module('lizard-nxt')
  */
 angular.module('lizard-nxt')
   .constant('versioning', window.versioning);
+
+/**
+ * @name production backend
+ * @memberOf app
+ * @description subdomain of production backend.
+ */
+angular.module('lizard-nxt')
+  .constant('backendDomain', 'https://demo.lizard.net/');